requirements.txt
click==7.1.2
contextlib2==0.6.0.post1
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
schema==0.7.4
SQLAlchemy==1.3.23
Werkzeug==1.0.1
pyjwt==2.0.1